Header Component Complete and Main Page Began
The header component has been completed with an aesthetic layout, functioning well on both chrome and microsoft edge browsers.
Added content for main page and organisation of this has begun. Next step to add images to main page and add content for other pages, and organise the structure of this further.